upgrade to 0.5.2

+++ Ilmn.Das.App.Wittyer/Vcf/Samples/ICopyNumberProvider.cs
@@ -8,6 +8,6 @@
         /// <value>
         /// The cn.
         /// </value>
-        uint? Cn { get; }
+        decimal? Cn { get; }
     }
 }
